Feature/tld ssl (#3)
* fix: Update check condition for existing lines in config file

* feat: Add script to configure reverse proxy with SSL

Add a bash script that prompts user for domain name and port to generate SSL certificate, configure Nginx for reverse proxy, and enable SSL.

* feat: Add Nginx installation to reverse-proxy script

Install Nginx during the reverse-proxy configuration setup to streamline the process.

* feat: Added logging messages to reverse-proxy.sh script

* feat: Improved error handling in reverse-proxy script

* feat: Added Tailscale domain configuration

Added functionality to generate a Tailscale domain and obtain a Tailscale certificate and key for secure access via HTTPS. Modified Nginx configuration to redirect both the main and Tailscale domains from HTTP to HTTPS.

* feat: Update Nginx configuration for Tailscale domain redirection

Included additional server block to redirect HTTP to HTTPS for Tailscale domain.
Cleaned up old configurations and generated self-signed SSL certificate and key.

* feat: Add automated domain name selection in reverse-proxy setup

Get the system hostname and allow user to use it as the default domain name in the reverse-proxy setup. Add logic to set a default port if not specified by the user during setup. Also, silence the output when generating dhparam.pem for SSL security.

* fix: Update hostname retrieval in reverse-proxy.sh

Update the hostname retrieval in the reverse-proxy.sh script to remove the use of the -f flag, which was causing issues on certain systems. Now, the script simply uses the 'hostname' command to get the hostname instead.